Fixed drop last known location issue - 2 + Rounded temp in domain models

+++ feature/forecast/src/main/java/com/iiddd/weather/forecast/domain/model/Weather.kt
@@ -1,7 +1,7 @@
 package com.iiddd.weather.forecast.domain.model
 
 data class Weather(
-    val currentTemp: Double,
+    val currentTemp: Int,
     val description: String,
     val hourly: List<HourlyForecast> = emptyList(),
     val daily: List<DailyForecast> = emptyList(),
@@ -10,13 +10,13 @@ data class Weather(
 
 data class HourlyForecast(
     val time: String,
-    val temp: Double,
+    val temp: Int,
     val icon: String
 )
 
 data class DailyForecast(
     val day: String,
-    val tempDay: Double,
-    val tempNight: Double,
+    val tempDay: Int,
+    val tempNight: Int,
     val icon: String
 )
